% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BotanicalCorrection.R
\name{BotanicalCorrection}
\alias{BotanicalCorrection}
\title{Botanical Correction}
\usage{
BotanicalCorrection(Data, Sources = c("tropicos", "usda", "wfo", "wcvp"))
}
\arguments{
\item{Data}{Dataset (data.frame or data.table); it must contain Site and IdTree}

\item{Sources}{Character vector. Taxonomic source(s) to use. Options are c("tropicos", "usda", "wfo", "wcvp") and default is all of them.}
}
\value{
A list with a complete log of the botanical corrections and Data with new columns:
\itemize{
\item \code{Accepted_family_DataHarmonizationCor} (character): corrected family name
\item \code{Accepted_genus_DataHarmonizationCor} (character): corrected genus name
\item \code{Accepted_species_DataHarmonizationCor} (character): corrected species name
\item \code{Accepted_name_DataHarmonizationCor} (character): corrected name (at any taxonomic level)
\item \code{Accepted_name_rank_DataHarmonizationCor} (character): Taxonomic rank of name matched by the TNRS
}
}
\description{
Botanical Correction
}
\details{
\itemize{
\item The function uses TNRS package provided by BIEN, https://bien.nceas.ucsb.edu/bien/tools/tnrs/
\item No special characters (typography)
\item The suffix "aceae" is restricted to families, and words ending with "aceae" are deleted anywhere else
\item Correct spelling of botanical names
\item Update the families according to APG
\item The log contains a flag if there is more than one name for one Site x IdTree combination
\item The log contains all details returned by TNRS (authors, etc.). Data just keeps the basic corrected names.
}
}
\examples{
\dontrun{
library(data.table)
data(TestData)
Rslt <- BotanicalCorrection(TestData)
}


}
