% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BotanicalCorrectionPlot.R
\name{BotanicalCorrectionPlot}
\alias{BotanicalCorrectionPlot}
\title{Plot Botanical correction results}
\usage{
BotanicalCorrectionPlot(Data)
}
\arguments{
\item{Data}{Dataset (data.frame or data.table)
The dataset must contain the columns:
\itemize{
\item \code{IdTree} (character)
\item \code{ScientificName} (character)
\item \code{ScientificName_DataHarmonizationCor} (character)
}}
}
\value{
Table of corrected species IDs
}
\description{
Plot Botanical correction results
}
\examples{

\dontrun{
data(TestData)
Rslt <- BotanicalCorrection(TestData)$Data
BotanicalCorrectionPlot(Rslt)
}

}
