% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataHarmonization-package.R
\docType{package}
\name{DataHarmonization}
\alias{DataHarmonization-package}
\alias{DataHarmonization}
\title{DataHarmonization-package}
\description{
Forest Inventories Harmonization & Correction
}
\section{DataHarmonization functions}{

RequiredFormat
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Alliance-for-Tropical-Forest-Science/DataHarmonization.git}
  \item Report bugs at \url{https://github.com/Alliance-for-Tropical-Forest-Science/DataHarmonization/issues}
}

}
\author{
\strong{Maintainer}: Vincyane Badouard \email{vincyane.badouard@gmail.com}

Authors:
\itemize{
  \item Valentine Herrmann \email{HerrmannV@si.edu}
  \item Nino Page \email{page.nino@gmail.com}
  \item Camille Piponiot \email{camille.piponiot@gmail.com}
}

}
\keyword{internal}
