% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiameterCorrection.R
\name{DiameterCorrection}
\alias{DiameterCorrection}
\title{Diameter correction}
\usage{
DiameterCorrection(
  Data,
  UseTaperCorrection = TRUE,
  DefaultHOM = 1.3,
  TaperParameter = function(DAB, HOM) 0.156 - 0.023 * log(DAB) - 0.021 * log(HOM),
  TaperFormula = function(DAB, HOM, TaperParameter, DefaultHOM) DAB/(exp(-TaperParameter
    * (HOM - DefaultHOM))),
  KeepMeas = c("MaxHOM", "MaxDate"),
  MinDBH = NULL,
  AddMissedRecruits = TRUE,
  AddMissedStems = TRUE,
  PositiveGrowthThreshold = 5,
  NegativeGrowthThreshold = -2,
  Pioneers = NULL,
  PioneersGrowthThreshold = 7.5,
  WhatToCorrect = c("POM change", "Abnormal growth"),
  CorrectionType = c("individual", "phylogenetic hierarchical"),
  DBHRange = 10,
  MinIndividualNbr = 5,
  DBHCorForDeadTrees = TRUE,
  coef = 0.9
)
}
\arguments{
\item{Data}{Dataset (data.frame or data.table)
The dataset must contain the columns:
\itemize{
\item \code{IdTree} or \code{IdStem} (character)
\item \code{ScientificName_DataHarmonizationCor} (character)
\item \code{Diameter} (numeric)
\item \code{IdCensus} (ordered factor)
\item \strong{\code{POM} (Point Of Measurement) (factor)} or
\strong{\code{HOM} (Height Of Measurement) (numeric)} if you want to correct from
the \strong{"POM change"}
If you want to apply the \strong{"phylogenetic hierarchical"} correction, the
dataset must also contain the columns:
\item \code{Genus_DataHarmonizationCor} (character)
\item \code{Family_DataHarmonizationCor} (character)
}}

\item{UseTaperCorrection}{(logical) TRUE: transform the tree diameter measured at a given height
into the diameter corresponding to the default measurement height (\code{DefaultHOM}), using an allometry.
FALSE: do not apply a taper correction}

\item{DefaultHOM}{Default Height Of Measurement in meter (Default: 1.3 m)
(numeric, 1 value)}

\item{TaperParameter}{Taper parameter (unitless) formula (function)
Default: \emph{TaperParameter = 0.156 - 0.023 log(DAB) - 0.021 log(HOM)}
of Cushman et al.2021.
With:
\itemize{
\item \emph{DAB}: Diameter Above Buttress (in cm)
\item \emph{HOM}: Height Of Measurement (in m)
}}

\item{TaperFormula}{Taper formula (function)
Default: \emph{DAB / (e^(- TaperParameter (HOM - DefaultHOM)))}
of Cushman et al.2021.
With:
\itemize{
\item \emph{DAB}: Diameter Above Buttress (in cm)
\item \emph{HOM}: Height Of Measurement (in m)
\item \emph{DefaultHOM}:  Default Height Of Measurement (in m)
\item \emph{TaperParameter}: Taper parameter (unitless)
}}

\item{KeepMeas}{In case of \strong{multiple diameter measurements} in the same
census:
Possible values: "MaxHOM", "MaxDate" (character).
\itemize{
\item "MaxHOM": apply the correction to the measurement taken at the
\strong{highest POM}
\item "MaxDate": apply the correction to the \strong{most recent measurement} (same
IdCensus but more recent date)
}}

\item{MinDBH}{Minimum diameter of trees inventoried (in cm) (numeric, 1 value) or
NULL (Default) if you wish to use the MinDBH indicated in your data, which may vary per plot}

\item{AddMissedRecruits}{(logical) TRUE: adds rows for stem that were supposed
to be recruited at a prior census, based on their estimated diameter (from linear regression)
and MinDBH. FALSE: will only indicate in the comment that the stem was supposed to be
recruited earlier.}

\item{AddMissedStems}{(logical) if TRUE, adds rows for trees that were missed between
two censuses, with their estimated diameter (based on linear regression)}

\item{PositiveGrowthThreshold}{in cm/year: a tree
widening by more than this value is considered abnormal (numeric, 1 value)}

\item{NegativeGrowthThreshold}{in cm/census: the possible
positive measurement error (+n) cannot be corrected until the growth
appears abnormal, but a negative measurement error can be allowed until -n
(a tree does not decrease). Thus the positive measurement error (+n) is
"compensated". (numeric, 1 value)}

\item{Pioneers}{Scientific names of the pioneer species of the site, as in
the \code{ScientificName_DataHarmonizationCor} column (characters vector)}

\item{PioneersGrowthThreshold}{in cm/year: a tree of a pioneer species that
widens by more than this value is considered abnormal (numeric, 1 value)}

\item{WhatToCorrect}{Possible values: "POM change", "Abnormal growth"
(character). All are complementary and recommended.
\itemize{
\item "POM change": detect POM change in the column \code{POM} and correct the
Diameter values from it. (Ignored if taper correction is applied)
\item "Abnormal growth": detect if the growth is greater than PositiveGrowthThreshold ('PioneersGrowthThreshold' if species belongs to 'Pioneers')
or smaller than NegativeGrowthThreshold and correct it by \code{CorrectionType}
}}

\item{CorrectionType}{Possible values: "individual", "phylogenetic
hierarchical" (character, 1 value).
\itemize{
\item "individual": replace abnormal growth by interpolation from the
individual values.
\item "phylogenetic hierarchical": replace abnormal growth with the average
growth of other trees in the dataset, at the specific, genus, family
or stand level, within a DBH range of x cm (\emph{DBHRange} argument).
If the number of these trees < n (\emph{MinIndividualNbr} argument)
at the specific level, we switch to the genus level etc.
}}

\item{DBHRange}{DBH range in cm to take into account to select other trees in
the dataset to apply "phylogenetic hierarchical" correction (Default: 10
cm) (numeric, 1 value)}

\item{MinIndividualNbr}{Minimum number of individuals to take into account in
"phylogenetic hierarchical" correction (Default: 5) (numeric, 1 value)}

\item{DBHCorForDeadTrees}{(logical) TRUE: return DBHCor also for dead trees.
FALSE: do not return DBHCor for dead trees. In this case it is advisable to
have corrected the tree life status with the \emph{StatusCorrection()} function.}

\item{coef}{(numeric, 1 value) This is used in individual corrections, to calculate weight of the growths by temporal proximity}
}
\value{
Fill the \emph{Comment_DataHarmonization} column with error type information and add columns:
\itemize{
\item \emph{Diameter_DataHarmonizationCor}: corrected trees diameter at default HOM
\item \emph{DiameterCorrectionMeth_DataHarmonization} = "local linear regression","weighted
mean"/phylogenetic hierarchical("species"/"genus"/"family"/"stand")/
"shift realignment"/"Same value".
\item \emph{POM_DataHarmonizationCor} (factor): POM value at which the corrected diameters are proposed.
Corresponds to the 1st POM value at which the stem was measured.
\item \emph{HOM_DataHarmonizationCor} (numeric): HOM value at which the corrected diameters
are proposed. Corresponds to the 1st HOM value at which the stem was
measured.
\item \emph{MissedStem_DataHarmonizationCor} if AddMissedStems is TRUE (with value TRUE for new rows).
\item \emph{MissedRecruit_DataHarmonizationCor} if AddMissedRecruits is TRUE (with value TRUE for new rows).
}
}
\description{
Diameter correction
}
\details{
When there is only 1 \code{Diameter} value for a tree/stem,
\code{Diameter_DataHarmonizationCor} takes the original \code{Diameter} value. If this value
is 0 or > MaxDBH, \code{Diameter_DataHarmonizationCor} takes NA. Diameters not linked to
an IdTree/IdStem or to a Census IdCensus are not processed.
Punctual error correction only with linear regression and not quadratic,
because punctual errors are corrected from a local regression with the 2
framing values.
}
\examples{
# library(data.table)
data(TestData)

TestData$HOM[1:3] <- c(0.5,1.5,NA)
TestData$Diameter[21:23] <- c(31,91,14)
TestData <- TestData[!(IdStem \%in\% "100658_1_auto" & IdCensus \%in\% 2017), ]

Rslt <- DiameterCorrection(
 TestData,
  WhatToCorrect = c("POM change", "Abnormal growth"),
    CorrectionType = c("phylo"),
    MinIndividualNbr = 1)

DiameterCorrectionPlot(Rslt, OnlyCorrected = TRUE)

}
