% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiameterCorrectionPlot.R
\name{DiameterCorrectionPlot}
\alias{DiameterCorrectionPlot}
\title{Plot diameter correction result}
\usage{
DiameterCorrectionPlot(
  Data,
  OnlyCorrected = TRUE,
  CorCol = "Diameter_DataHarmonizationCor",
  SeveralWindows = TRUE
)
}
\arguments{
\item{Data}{Dataset (data.frame or data.table)
The dataset must contain the columns:
\itemize{
\item \code{IdStem} (character)
\item \code{IdCensus} (ordered factor)
\item \code{Diameter} (numeric)
\item \code{DBHCor} (numeric)
\item \code{HOM} (Height Of Measurement) (numeric)
\item \code{HOMCor} (Corrected Height Of Measurement) (numeric)
}}

\item{OnlyCorrected}{TRUE: plot only corrected stems, FALSE: plot all stems
(logical)}

\item{CorCol}{Corrected column name (character)}

\item{SeveralWindows}{TRUE: return each page in a new window (better
visualisation in Rstudio), FALSE: return each page in the same window
(needed to save all the pages) (logical)}
}
\value{
The plots of the initial diameter values and proposed corrections, by
IdStem.
}
\description{
Plot diameter correction result
}
\examples{

\dontrun{
# pdf("DiameterCorrectionPlots_TestData.pdf", width = 25, height = 10)

 data(TestData)

TestData$HOM[1:3] <- c(0.5,1.5,NA)
TestData$Diameter[1:3] <- TestData$Diameter[1:3] + c(2,-2,0)
TestData$Diameter[21:23] <- c(31,91,14)

Rslt <- DiameterCorrection(
 TestData,
  WhatToCorrect = c("POM change","Abnormal growth"),
   CorrectionType = c("phylo"),
    MinIndividualNbr = 1)
DiameterCorrectionPlot(Rslt, OnlyCorrected = TRUE, SeveralWindows = FALSE)

# dev.off()
}

}
