% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestGeoSubset.R
\docType{data}
\name{ForestGeoSubset}
\alias{ForestGeoSubset}
\title{ForestGeoSubset}
\format{
See examples.
\describe{
To interpret the names of the dataset, see \url{http://ctfs.si.edu/Public/CTFSRPackage/files/help/RoutputStem.pdf} and \url{https://raw.githubusercontent.com/forestgeo/fgeo.data/master/man/census_description.Rd}
...
}
}
\usage{
ForestGeoSubset
}
\description{
Stem tables from Luquillo produced with
\href{https://github.com/forestgeo/rtbl/releases/tag/0.0.0.9000}{\strong{rtbl} 0.0.0.9000}, with Species table merge into it and some alteration of dbh for troubleshooting purpous. NOT SUITABLE FOR DATA ANALYSIS.
}
\seealso{
\url{https://forestgeo.si.edu/sites/north-america/luquillo}.
}
\keyword{datasets}
