% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullErrorProcessing.R
\name{FullErrorProcessing}
\alias{FullErrorProcessing}
\title{Full error processing}
\usage{
FullErrorProcessing(
  Data,
  DeathConfirmation = 2,
  UseSize = FALSE,
  AddRowsForForgottenCensuses = TRUE,
  UseTaperCorrection = TRUE,
  DefaultHOM = 1.3,
  TaperParameter = function(DAB, HOM) 0.156 - 0.023 * log(DAB) - 0.021 * log(HOM),
  TaperFormula = function(DAB, HOM, TaperParameter, DefaultHOM) DAB/(exp(-TaperParameter
    * (HOM - DefaultHOM))),
  KeepMeas = c("MaxHOM", "MaxDate"),
  PositiveGrowthThreshold = 5,
  NegativeGrowthThreshold = -2,
  Pioneers = NULL,
  PioneersGrowthThreshold = 7.5,
  WhatToCorrect = c("POM change", "Abnormal growth"),
  CorrectionType = c("individual", "phylogenetic hierarchical"),
  DBHRange = 10,
  MinIndividualNbr = 5,
  MinDBH = NULL,
  AddMissedRecruits = T,
  AddMissedStems = T,
  DBHCorForDeadTrees = TRUE,
  coef = 0.9,
  OnlyDetectMissedRecruits = FALSE
)
}
\arguments{
\item{Data}{Dataset (data.frame or data.table)}

\item{DeathConfirmation}{Number of times (censuses) needed for an unseen tree
to be considered dead (numeric)  (Default = 2 censuses)}

\item{UseSize}{Use the size presence  (> min DBH) as a witness of the living status of the
tree (logical) (Default = FALSE)}

\item{AddRowsForForgottenCensuses}{TRUE: adds rows for forgotten censuses, FALSE: does not add any rows (logical).}

\item{UseTaperCorrection}{(logical) TRUE: transform the tree diameter measured at a given height
into the diameter corresponding to the default measurement height (\code{DefaultHOM}), using an allometry.
FALSE: do not apply a taper correction}

\item{DefaultHOM}{Default Height Of Measurement in meter (Default: 1.3 m)
(numeric, 1 value)}

\item{TaperParameter}{Taper parameter (unitless) formula (function)
Default: \emph{TaperParameter = 0.156 - 0.023 log(DAB) - 0.021 log(HOM)}
of Cushman et al.2021.
With:
\itemize{
\item \emph{DAB}: Diameter Above Buttress (in cm)
\item \emph{HOM}: Height Of Measurement (in m)
}}

\item{TaperFormula}{Taper formula (function)
Default: \emph{DAB / (e^(- TaperParameter (HOM - DefaultHOM)))}
of Cushman et al.2021.
With:
\itemize{
\item \emph{DAB}: Diameter Above Buttress (in cm)
\item \emph{HOM}: Height Of Measurement (in m)
\item \emph{DefaultHOM}:  Default Height Of Measurement (in m)
\item \emph{TaperParameter}: Taper parameter (unitless)
}}

\item{KeepMeas}{In case of \strong{multiple diameter measurements} in the same
census year:
Possible values: "MaxHOM", "MaxDate" (character).
\itemize{
\item "MaxHOM": apply the correction to the measurement taken at the
\strong{highest POM}
\item "MaxDate": apply the correction to the \strong{most recent measurement} (same
year but more recent date)
}}

\item{PositiveGrowthThreshold}{in cm/year: a tree
widening by more than this value is considered abnormal (numeric, 1 value)}

\item{NegativeGrowthThreshold}{in cm/census: the possible
positive measurement error (+n) cannot be corrected until the growth
appears abnormal, but a negative measurement error can be allowed until -n
(a tree does not decrease). Thus the positive measurement error (+n) is
"compensated". (numeric, 1 value)}

\item{Pioneers}{Scientific names of the pioneer species of the site, as in
the \code{ScientificName_DataHarmonizationCor} column (characters vector)}

\item{PioneersGrowthThreshold}{in cm/year: a tree of a pioneer species that
widens by more than this value is considered abnormal (numeric, 1 value)}

\item{WhatToCorrect}{Possible values: "POM change", "Abnormal growth"
(character). All are complementary and recommended.
\itemize{
\item "POM change": detect POM change in the column \code{POM} and correct the
Diameter values from it. (Ignored if taper correction is applied)
\item "Abnormal growth": detect if the growth is greater than PositiveGrowthThreshold ('PioneersGrowthThreshold' if species belongs to 'Pioneers')
or smaller than NegativeGrowthThreshold and correct it by \code{CorrectionType}
}}

\item{CorrectionType}{Possible values: "individual", "phylogenetic
hierarchical" (character, 1 value).
\itemize{
\item "individual": replace abnormal growth by interpolation from the
individual values.
\item "phylogenetic hierarchical": replace abnormal growth with the average
growth of other trees in the dataset, at the specific, genus, family
or stand level, within a DBH range of x cm (\emph{DBHRange} argument).
If the number of these trees < n (\emph{MinIndividualNbr} argument)
at the specific level, we switch to the genus level etc.
}}

\item{DBHRange}{DBH range in cm to take into account to select other trees in
the dataset to apply "phylogenetic hierarchical" correction (Default: 10
cm) (numeric, 1 value)}

\item{MinIndividualNbr}{Minimum number of individuals to take into account in
"phylogenetic hierarchical" correction (Default: 5) (numeric, 1 value)}

\item{MinDBH}{Minimum diameter of trees inventoried (in cm) (numeric, 1 value) or
NULL (Default) if you wish to use the MinDBH indicated in your data, which may vary per plot}

\item{AddMissedRecruits}{(logical) TRUE: adds rows for stem that were supposed
to be recruited at a prior census, based on their estimated diameter (from linear regression)
and MinDBH. FALSE: will only indicate in the comment that the stem was supposed to be
recruited earlier.}

\item{AddMissedStems}{(logical) if TRUE, adds rows for trees that were missed between
two censuses, with their estimated diameter (based on linear regression)}

\item{DBHCorForDeadTrees}{(logical) TRUE: return DBHCor also for dead trees.
FALSE: do not return DBHCor for dead trees. In this case it is advisable to
have corrected the tree life status with the \emph{StatusCorrection()} function.}

\item{coef}{(numeric, 1 value) This is used in individual corrections, to calculate weight of the growths by temporal proximity}

\item{OnlyDetectMissedRecruits}{TRUE: Only detect errors, FALSE: detect and correct errors
(logical)}
}
\value{
The original dataset (data.table) with a \emph{Comment} column containing
information on the errors detected per row, the correction columns, and
columns containing correction methods.
}
\description{
Full error processing
}
\details{
Detect errors or detect and correct errors:
\itemize{
\item Check general errors (\emph{GeneralErrorsDetection})
\item Check botanical identification (\emph{BotanicalCorrection})
\item Check the life status evolution of the trees/stems (\emph{StatusCorrection})
\item Apply a taper allometry on diameters measured at heights different from the
default HOM (\emph{TaperCorrection})
\item Check diameter evolution of the trees (\emph{DiameterCorrection})
\item Check tree/stem recruitment (\emph{RecruitmentCorrection})
}
}
\examples{

\dontrun{
data(TestData)
Rslt <- FullErrorProcessing(TestData,
 OnlyDetectMissedRecruits = TRUE,
  AddRowsForForgottenCensuses = FALSE)

Rslt_Test <- FullErrorProcessing(TestData)
Rslt_Panama <- FullErrorProcessing(PanamaFormated)
}

}
\seealso{
\link{GeneralErrorsDetection}, \link{BotanicalCorrection},
\link{StatusCorrection}, \link{TaperCorrection}, \link{DiameterCorrection},
\link{RecruitmentCorrection}
}
