% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralErrorsDetection.R
\name{GeneralErrorsDetection}
\alias{GeneralErrorsDetection}
\title{General Errors Detection}
\usage{
GeneralErrorsDetection(Data)
}
\arguments{
\item{Data}{Dataset (data.frame or data.table)}
}
\value{
The input dataset (data.table) with a new \emph{Comment} column with error
type informations.
}
\description{
General Errors Detection
}
\details{
Detect errors
\itemize{
\item Remove \strong{duplicated rows}
\item Check \strong{missing value} in
X-YTreeUTM/PlotArea/Plot/Subplot/Year/TreeFieldNum/
IdTree/IdStem/Diameter/POM/HOM/Family/Genus/Species/VernName
\item Check \strong{missing value} (NA/0) in the measurement variables: "Diameter",
"HOM", "TreeHeight", "StemHeight"
\item Check of the \strong{unique association of the IdTree with plot, subplot}
\strong{and TreeFieldNum} (at the site scale)
\item Check \strong{duplicated IdTree/IdStem} in a census (at the site scale)
\item Check for trees \strong{outside the subplot} (not implemented yet)
\item Check \strong{invariant coordinates per IdTree/IdStem}
\item Check \strong{fix Plot and Subplot number} (not implemented yet)
}
}
\examples{
library(data.table)
data("TestData")

Rslt <- GeneralErrorsDetection(TestData)

}
