% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateComment.R
\name{GenerateComment}
\alias{GenerateComment}
\title{GenerateComment}
\usage{
GenerateComment(x, comment)
}
\arguments{
\item{x}{a character vector}

\item{comment}{The string to add in the column (character)}
}
\value{
concatenation of x and comment, separated by a semicolon
}
\description{
Helper function to concatenate comments, separated by ";"
}
\examples{
GenerateComment(letters, "A")
}
