% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParacouSubset.R
\docType{data}
\name{ParacouSubset}
\alias{ParacouSubset}
\title{ParacouSubset}
\format{
A tibble with 4854 rows and 31 variables:
\describe{
\item{Forest}{Forest name (character)}
\item{Plot}{Plot number (character)}
\item{PlotArea}{Plot area, in hectare (numeric)}
\item{SubPlot}{Subplot number (integer)}
\item{SubSubPlot}{Dummy sub-sub-plots created for the package (integer)}
\item{TreeFieldNum}{Tree unique identifiers within the subplot (matching
the tag number in the field) (numeric)}
\item{idTree}{Unique tree identifier from the database (integer)}
\item{Protocole}{Protocole name (character)}
\item{Xfield}{Euclidean position of the tree relative to
the Southwestern corner of the subplot, in m (numeric)}
\item{Yfield}{Euclidean position of the tree relative to
the Southwestern corner of the subplot, in m (numeric)}
\item{Xutm}{Coordinate X for the tree in UTM 22 N (EPSG: 32 622) (numeric)}
\item{Yutm}{Coordinate Y for the tree in UTM 22 N (EPSG: 32 622) (numeric)}
\item{UTMZone}{UTM Zone (integer)}
\item{Lat}{Tree latitude (WGS 84, EPSG : 4326) (numeric)}
\item{Lon}{Tree longitude (WGS 84, EPSG : 4326) (numeric)}
\item{Family}{Botanical family (character)}
\item{Genus}{Botanical genus (character)}
\item{Species}{Botanical species (character)}
\item{BotaSource}{Source of botanical name (character)}
\item{BotaCertainty}{Level of certainty
for the botanical identification (numeric)}
\item{idVern}{Unique vernacular identifier from the database (numeric)}
\item{VernName}{Vernacular name (character)}
\item{CommercialSp}{Is the tree considered as a commercial species
according to the list of species that were logged during the sylvicultural
treatment in Paracou (TRUE) or not (FALSE) ? (logical)}
\item{CensusYear}{Census year (integer)}
\item{CensusDate}{Census Day (when available) (character)}
\item{CensusDateCertainty}{Is the CensusDate precise (TRUE) or not (FALSE)?
If FALSE, this information must not be considered for calculation (logical)}
\item{CodeAlive}{Is the tree alive (TRUE) or dead (FALSE)? (logical)}
\item{MeasCode}{Information on the method for measuring the circumference
or on the state of the tree (integer)}
\item{Circ}{Circumference of the tree at 1.30m above ground (DBH level),
in cm (numeric)}
\item{CircCorr}{Corrected circumference
(mathematical correction of abnormal circumferences), in cm (numeric)}
\item{CorrCode}{Information on the reason why and how the circumference has
been corrected, multiple values are possible(character)}
...
}
}
\source{
\url{http://paracou.cirad.fr} ;
\url{https://paracoudata.cirad.fr/public/pdf/Paracou_data_dictionnary.pdf}
}
\usage{
ParacouSubset
}
\description{
Subset of tree inventories of Paracou research station (French Guiana)
\itemize{
\item Plot 6
\item SubPlot 1
\item 2016-2020
\item 5 last censuses
\item 1000 random individuals
\item 4 dummy sub-sub-squares created for the package to be use as plots
in long format.
Dataset extracted from the Guyafor database.
}
}
\keyword{datasets}
