% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecruitmentCorrection.R
\name{RecruitmentCorrection}
\alias{RecruitmentCorrection}
\title{RecruitmentCorrection}
\usage{
RecruitmentCorrection(
  Data,
  KeepMeas = c("MaxHOM", "MaxDate"),
  MinDBH = NULL,
  OnlyDetectMissedRecruits = FALSE
)
}
\arguments{
\item{Data}{(data.frame or data.table)
The dataset should preferably contain the column of corrected diameters:
'DBH_DataHarmonizationCor', otherwise the function will take the column 'Diameter'}

\item{KeepMeas}{In case of \strong{multiple diameter measurements} in the same
census year, on which to apply the correction:
Possible values: "MaxHOM", "MaxDate" (character).
\itemize{
\item "MaxHOM": apply the correction to the measurement taken at the
\strong{highest POM}
\item "MaxDate": apply the correction to the \strong{most recent measurement} (same
year but more recent date)
}}

\item{MinDBH}{Minimum diameter of trees inventoried (in cm) (numeric, 1 value) or
NULL (Default) if you wish to use the MinDBH indicated in your data, which may vary per plot}

\item{OnlyDetectMissedRecruits}{TRUE: Only detect errors, FALSE: detect and correct errors
(logical)}
}
\value{
When non in OnlyDetectMissedRecruits mode, add rows for forgotten recruits with their estimated DBH in the
'Diameter_DataHarmonizationCor' column, create a 'CorrectedRecruit' col (logical) to indicate
the rows that were added.
In OnlyDetectMissedRecruits mode, fill the 'Comment' column : "This DBH is/was the 1st recorded for this
tree, according to its annual growth and the census done for this plot, it
should have been recruited earlier according to your protocol (MinDBH)."
}
\description{
RecruitmentCorrection
}
\details{
This function detects trees that may have been missed
at a previous census if the estimated diameter at that census is
greater than MinDBH. Estimated diameter is calculated
by recursively removing the tree's first growth measurement to it's first
diameter measurement. If growth can't be calculated for a tree, we use
the average first growth of trees of the same species, if not possible,
of the same genus, and if not possible, of all trees.
It is strongly recommended to correct the Diameter before correcting the recruits.
}
\examples{
library(data.table)
data(TestData)
setnames(TestData, "Diameter", "Diameter_DataHarmonizationCor", skip_absent=TRUE)

Rslt <- RecruitmentCorrection(TestData)
IdCorr <- Rslt[CorrectedRecruit \%in\%  TRUE, IdStem]
TreesCorr <- Rslt[IdStem \%in\% IdCorr]

library(ggplot2)
ggplot(TreesCorr) +
aes(x = Year, y = Diameter_DataHarmonizationCor) +
  geom_line(size = 0.5, colour = "#112446") +
  geom_point(shape = "circle", size = 1.5, mapping = aes(color = CorrectedRecruit)) +
  theme_minimal() +
  facet_wrap(vars(IdStem), scales = "free")

}
