% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReversedRequiredFormat.R
\name{ReversedRequiredFormat}
\alias{ReversedRequiredFormat}
\title{ReversedRequiredFormat}
\usage{
ReversedRequiredFormat(Data, input, x = NULL, Untidy = FALSE)
}
\arguments{
\item{Data}{Standardized data, returned from \code{\link[=RequiredFormat]{RequiredFormat()}}}

\item{input}{A named list, typically the output of function
RequiredFormat_interactive, also called site profile with information on
column names correspondence, size units etc... Chosen to be the output
profile the user wants their data turned into}

\item{x}{For internal use when function used by Shiny app}

\item{Untidy}{(logical). If TRUE and input$tidy exists, the data will be
untidy (changed from long to wide format, according to input information)}
}
\value{
(data.frame) in the format given the profile selected in input.
}
\description{
ReversedRequiredFormat
}
\details{
This function takes the standardized forest inventory data.table
(returned by \code{\link[=RequiredFormat]{RequiredFormat()}}) and converts the column names to the names
of the profile given as input.
}
\examples{
\dontrun{

data(ParacouSubsetFormated)
data("ForestGeoProfile")
ReversedRequiredFormat(ParacouSubsetFormated, ForestGeoProfile)
               }

}
