% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCBISubsetFormated.R
\docType{data}
\name{SCBISubsetFormated}
\alias{SCBISubsetFormated}
\title{SCBISubsetFormated}
\format{
A tibble
\describe{
\item{MinDBH}{The minimum diameter of trees included in the inventory}
\item{IdCensus}{Census identifier}
\item{Year}{Year of measurement}
\item{Month}{Month of measurement}
\item{Day}{Day of measurement}
\item{Date, DateOriginal}{Date of measurement}
\item{Site}{Site level name}
\item{Cluster}{Cluster ID or name (This was created to accommodate ForestPlots network's cluster, which groups similar plots and is used by their data packages to create local allometries. Note that it is useful to also keep Site (PlotName) so as users able to check for duplicate datasets from multiple networks).}
\item{Plot}{Plot level name}
\item{PlotViewID}{additional plot protocol identifier useful to indicate if the plot protocol has changed and/or different soil type/disturbance... (This was created to accommodate ForestPlots network).}
\item{Subplot}{Subplot level name}
\item{PlotArea}{Size of the plot}
\item{PlotElevation}{plot's elevation in meters above sea level}
\item{SubplotArea}{Size of the plot}
\item{PlotLat}{Plot's latitude in degrees}
\item{PlotLon}{Plot's longitude in degrees}
\item{XPlotUTM}{Plot's X UTM coordinates}
\item{YPlotUTM}{Plot's Y UTM coordinates}
\item{SubplotLat}{Subplot's latitude in degrees}
\item{SubplotLon}{Subplot's longitude in degrees}
\item{XSubplotUTM}{Subplot's X UTM coordinates}
\item{YSubplotUTM}{Subplot's Y UTM coordinates}
\item{ScientificName, ScientificNameOriginal}{Scientific name}
\item{VernName}{Vernacular name}
\item{Family}{Family name}
\item{Genus}{Genus name}
\item{Species}{Species name}
\item{Subspecies}{Subspecies name}
\item{Variety}{Variety name}
\item{Voucher}{Voucher code for the species}
\item{IdLevel}{Deepest taxonomic level for which full identification is known}
\item{Authority}{Taxonomic authority for the ID level}
\item{CommercialSp}{Logical: TRUE if species is considered commercial, FALSE if not}
\item{LifeForm}{Life form}
\item{TreeFieldNum, TreeFieldNumOriginal}{Tree unique identifiers, matching the tag number in the field.}
\item{IdTree, IdTreeOriginal}{Tree unique identifiers. This was automatically generated if it was not provided in the input data (using plot and tree tag information).}
\item{StemFieldNum, StemFieldNumOriginal}{Stem identifiers within the tree.}
\item{IdStem, IdStemOriginal}{Stem unique identifiers. This was automatically generated if it was not provided in the input data (using plot tree and stem tag information).}
\item{IdMeasure}{measure unique identifier if you had more than one measure per stem per year}
\item{TreeLat}{Tree Latitude in degrees}
\item{TreeLon}{Tree Longitude in degrees}
\item{XTreeUTM}{Tree X UTM coordinates}
\item{YTreeUTM}{Tree Y UTM coordinates}
\item{XTreePlot}{Tree X euclidean position in plot}
\item{YTreePlot}{Tree Y euclidean position in plot}
\item{XTreeSubplot}{Tree X euclidean position in subplot}
\item{YTreeSubplot}{Tree Y euclidean position in subplot}
\item{LifeStatus, LifeStatusOriginal}{Logical: TRUE if tree was alive, FALSE if it was dead}
\item{Diameter}{Tree Diameter}
\item{BD}{Basal diameter}
\item{Circ}{Circumference}
\item{BCirc}{Basal circumference}
\item{HOM}{Height of measurement}
\item{POM}{Code for point of measurement of tree diameter}
\item{BHOM}{Height of measurement of basal diameter}
\item{BPOM}{Code for point of measurement of basal diameter}
\item{TreeHeight}{Tree Height}
...
}
}
\usage{
SCBISubsetFormated
}
\description{
This is the standardized version of a subset of Smithsonian Conservation Biology Institute's ForestGEO plot's data, for which some errors were introduced to test the correction functions
}
\keyword{datasets}
