% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatusCorrectionPlot.R
\name{StatusCorrectionPlot}
\alias{StatusCorrectionPlot}
\title{Plot life status correction result}
\usage{
StatusCorrectionPlot(Data, OnlyCorrected = TRUE, SeveralWindows = TRUE)
}
\arguments{
\item{Data}{Dataset (data.frame or data.table)
The dataset must contain the columns:
\itemize{
\item \code{IdTree} (character)
\item \code{Year} (numeric)
\item \code{LifeStatus} (logical or character)
\item \code{LifeStatus_DataHarmonizationCor} (logical or character)
}}

\item{OnlyCorrected}{TRUE: plot only corrected trees, FALSE: plot all trees
(logical)}

\item{SeveralWindows}{TRUE: return each page in a new window (better
visualisation in Rstudio), FALSE: return each page in the same window
(needed to save all the pages) (logical)}
}
\value{
The plots of the initial diameter values and proposed corrections, by
IdStem/IdTree
}
\description{
Plot life status correction result
}
\examples{

\dontrun{
data(TestData)
Rslt <- StatusCorrection(TestData)
StatusCorrectionPlot(Rslt, SeveralWindows = TRUE)
}

}
