% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaperCorrection.R
\name{TaperCorrection}
\alias{TaperCorrection}
\title{TaperCorrection}
\usage{
TaperCorrection(
  Data,
  DefaultHOM = 1.3,
  TaperParameter = function(DAB, HOM) 0.156 - 0.023 * log(DAB) - 0.021 * log(HOM),
  TaperFormula = function(DAB, HOM, TaperParameter, DefaultHOM) DAB/(exp(-TaperParameter
    * (HOM - DefaultHOM)))
)
}
\arguments{
\item{Data}{Dataset (data.frame or data.table)
The dataset must contain the columns:
\itemize{
\item \code{Diameter} (numeric)
\item \strong{\code{HOM} (Height Of Measurement) (numeric)}
}}

\item{DefaultHOM}{Default Height Of Measurement in meter (Default: 1.3 m)
(numeric, 1 value)}

\item{TaperParameter}{Taper parameter (unitless) formula (function)
Default: \emph{TaperParameter = 0.156 - 0.023 log(DAB) - 0.021 log(HOM)}
of Cushman et al.2021.
With:
\itemize{
\item \emph{DAB}: Diameter Above Buttress (in cm)
\item \emph{HOM}: Height Of Measurement (in m)
}}

\item{TaperFormula}{Taper formula (function)
Default: \emph{DAB / (e^(- TaperParameter (HOM - DefaultHOM)))}
of Cushman et al.2021.
With:
\itemize{
\item \emph{DAB}: Diameter Above Buttress (in cm)
\item \emph{HOM}: Height Of Measurement (in m)
\item \emph{DefaultHOM}:  Default Height Of Measurement (in m)
\item \emph{TaperParameter}: Taper parameter (unitless)
}}
}
\value{
Fill the \emph{Comment_DataHarmonization} column with error type informations and add columns:
- \emph{Diameter_DataHarmonizationCor}: corrected trees diameter at default HOM
- \emph{DiameterCorrectionMeth} = "taper"
- \emph{HOM_DataHarmonizationCor} (numeric): HOM corresponding to the
\emph{Diameter_DataHarmonizationCor} (= \emph{DefaultHOM})
}
\description{
Transform the tree diameter measured at a given height into the
diameter corresponding to the default measurement height (HOM), using an
allometry.
}
\examples{
library(data.table)

Data <- data.table(IdStem = "A",
      ScientificName = "Tree",
      Year = c(1998, 2008, 2016, 2017, 2018, 2019, 2021),
      IdCensus = factor(c(1998, 2008, 2016, 2017, 2018, 2019, 2021), ordered = TRUE),
      Diameter = c(19, 19, 21.4, 22.6, 23.1, 23.1, 23.6),
      HOM = c(1.30, 3.25, 3.25, 3.25, 3.25, 3.25, 3.25))

Rslt <- TaperCorrection(Data)
DiameterCorrectionPlot(Rslt, CorCol = "Diameter_DataHarmonizationCor")

}
