% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestData.R
\docType{data}
\name{TestData}
\alias{TestData}
\title{Test Data of the DataHarmonization package}
\format{
A tibble with 4855 rows and 32 variables:
\describe{
\item{Site}{Site name (character)}
\item{Plot}{Plot number (character)}
\item{PlotArea}{Plot area, in hectare (numeric)}
\item{Subplot}{Subplot number (character)}
\item{SubSubplot}{Dummy sub-sub-plots created for the package (character)}
\item{TreeFieldNum}{Tree unique identifiers within the subplot (matching
the tag number in the field) (character)}
\item{IdTree}{Unique tree identifier from the database (character)}
\item{Protocole}{Protocole name (character)}
\item{XTreePlot}{Euclidean position of the tree relative to
the Southwestern corner of the subplot, in m (numeric)}
\item{YTreePlot}{Euclidean position of the tree relative to
the Southwestern corner of the subplot, in m (numeric)}
\item{XTreeUTM}{Coordinate X for the tree in UTM 22 N (EPSG: 32 622) (numeric)}
\item{YTreeUTM}{Coordinate Y for the tree in UTM 22 N (EPSG: 32 622) (numeric)}
\item{UTMZone}{UTM Zone (integer)}
\item{TreeLat}{Tree latitude (WGS 84, EPSG : 4326) (numeric)}
\item{TreeLon}{Tree longitude (WGS 84, EPSG : 4326) (numeric)}
\item{Family}{Botanical family (character)}
\item{Genus}{Botanical genus (character)}
\item{Species}{Botanical species (character)}
\item{ScientificName}{Scientific name (character)}
\item{BotaSource}{Source of botanical name (character)}
\item{BotaCertainty}{Level of certainty
for the botanical identification (numeric)}
\item{idVern}{Unique vernacular identifier from the database (numeric)}
\item{VernName}{Vernacular name (character)}
\item{CommercialSp}{Is the tree considered as a commercial species
according to the list of species that were logged during the sylvicultural
treatment in Paracou (TRUE) or not (FALSE) ? (logical)}
\item{Year}{Census year (numeric)}
\item{Date}{Census Day (when available) (character)}
\item{DateCertainty}{Is the Date precise (TRUE) or not (FALSE)?
If FALSE, this information must not be considered for calculation (logical)}
\item{LifeStatus}{Is the tree alive (TRUE) or dead (FALSE)? (logical)}
\item{POM}{Point Of Measurement as a code (integer)}
\item{Diameter}{Tree diameter in cm (numeric)}
\item{CircCorr}{Corrected circumference
(mathematical correction of abnormal circumferences), in cm (numeric)}
\item{CorrCode}{Information on the reason why and how the circumference has
been corrected, multiple values are possible(character)}
...
}
}
\source{
\url{http://paracou.cirad.fr} ;
\url{https://paracoudata.cirad.fr/public/pdf/Paracou_data_dictionnary.pdf}
}
\usage{
TestData
}
\description{
\emph{Attention}: this inventory contains errors that have been deliberately added
in order to test their detection.
}
\details{
Subset of tree inventories of Paracou research station (French Guiana)
\itemize{
\item Plot 6
\item SubPlot 1
\item 2016-2020
\item 5 last censuses
\item 1000 random individuals
\item 4 dummy sub-sub-squares created for the package to be use as plots
in long format.
Dataset extracted from the Guyafor database \emph{WITH ADDING ERRORS}
}
}
\keyword{datasets}
